requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
billiard==4.2.4
cachetools==6.2.4
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
distro==1.9.0
Django==5.0
djangorestframework==3.14.0
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.45.0
google-genai==1.56.0
google-generativeai==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.2
pytest==7.4.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==7.1.0
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.5.0
websockets==15.0.1
